Let New Project pick a Forms theme; share the picker UI with Add Forms
New Project always imported the default theme with no way to choose. Extract
AddFormsViewModel's theme picker (available themes, selection, project-change
requirements panel) into ThemeSelectionViewModel plus a shared
ThemeSelectionControl, and compose both into NewProjectDialogViewModel and
AddFormsViewModel. Sets up the planned theme-preview PR to touch one control
instead of two.
