+++ Gum/Dialogs/NewProjectDialogView.xaml
@@ -1,18 +1,23 @@
-﻿<UserControl
+<UserControl
     x:Class="Gum.Dialogs.NewProjectDialogView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:controls="clr-namespace:Gum.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:dialogs="clr-namespace:Gum.Services.Dialogs"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    x:Name="Root"
     dialogs:Dialog.DialogTitle="New Project"
     mc:Ignorable="d">
     <StackPanel MaxWidth="420" Orientation="Vertical">
         <CheckBox IsChecked="{Binding IsIncludeFormsControls}">Include Forms controls</CheckBox>
         <TextBlock
-            Margin="0,4,0,0"
+            Margin="0,4,0,8"
             Foreground="{DynamicResource Frb.Brushes.Foreground.Subtle}"
             Text="Adds buttons, text boxes, list boxes and other standard controls to the project."
             TextWrapping="Wrap" />
+        <controls:ThemeSelectionControl
+            DataContext="{Binding ThemeSelection}"
+            IsEnabled="{Binding DataContext.IsIncludeFormsControls, ElementName=Root}" />
     </StackPanel>
 </UserControl>
